Add files via upload
An important part of business is planning for the future and ensuring that the company survives changing market conditions. Some businesses do this really well and last for hundreds of years. This project takes a look at data from BusinessFinancing.co.uk on the world's oldest businesses: when they were founded and which industries they belong to.