Add files via upload
This notebook contains data provided by the United States Social Security Administration, which lists first names along with the number and sex of babies they were given to in each year. The dataset was limited to first names which were given to over 5,000 American babies in a given year. This data spans 101 years, from 1920 through 2020.